src/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions==1.12.0
pycodestyle==2.10.0
pytest==7.1.3
azure-identity==1.12.0
azure-mgmt-storage==21.0.0
azure-storage-file-datalake==12.9.1
pydantic==1.8.2
openai
